# Optimizing trimesh by implementing vertex-to-poly lookup and reducing some iterators