I made another set of micro optimizations, this time gains are a little bit larger (mainly in trimesh optimizations)

In `tests-aurora` arround 10% speedup