@PTFOPlayer 2025-01-18
The clippy warning has been fixed, and the spade version has been reverted to 2.10.0