@mockersf 2025-01-20
I ran benchmarks on this PR, and notice a 10% perf degradation on the "get path" benches

```
group                               main                     this PR
-----                               -------                  -------
get path Vec2(233.0, 323.0)         1.00   948.8±18.79µs     1.10  1039.8±27.49µs
get path Vec2(356.0, 166.0)         1.00   969.2±20.09µs     1.11  1073.1±25.34µs
get path Vec2(468.0, 584.0)         1.00   655.6±18.27µs     1.10   721.1±18.19µs
get path Vec2(512.0, 170.0)         1.00   853.6±16.60µs     1.09   934.0±21.06µs
get path Vec2(611.0, 658.0)         1.00   982.4±22.58µs     1.10  1079.6±26.83µs
get path Vec2(827.0, 678.0)         1.00  1024.0±35.47µs     1.10  1122.9±26.46µs
get path Vec2(993.0, 290.0)         1.00      2.1±0.04ms     1.10      2.3±0.02ms
```

Could you check if you get similar results?

You should be able to run them with `cargo bench "get path"`, and I used [critcmp](https://github.com/BurntSushi/critcmp) for the comparison