@PTFOPlayer 2025-01-20
It seems like there is some regression when vertex-to-poly lookup is beeing done.

I ran benchmarks with and those are results:

With lookup:
```
group                          after                                  before
-----                          -----                                  ------
get path Vec2(233.0, 323.0)    1.00  1901.4±24.07µs        ? ?/sec    3.65      6.9±7.92ms        ? ?/sec
get path Vec2(356.0, 166.0)    1.19      2.1±0.03ms        ? ?/sec    1.00  1736.4±42.28µs        ? ?/sec
get path Vec2(468.0, 584.0)    1.00  1415.5±20.93µs        ? ?/sec    9.72     13.8±2.02ms        ? ?/sec
get path Vec2(512.0, 170.0)    1.00  1757.6±23.33µs        ? ?/sec    1.45      2.5±1.00ms        ? ?/sec
get path Vec2(611.0, 658.0)    1.00  1992.7±14.58µs        ? ?/sec    1.26      2.5±0.70ms        ? ?/sec
get path Vec2(827.0, 678.0)    1.14      2.2±0.04ms        ? ?/sec    1.00  1909.9±35.95µs        ? ?/sec
get path Vec2(993.0, 290.0)    1.08      3.6±0.11ms        ? ?/sec    1.00      3.3±0.02ms        ? ?/sec
```

Without lookup:

```
group                          before                                 no_hash
-----                          ------                                 -------
get path Vec2(233.0, 323.0)    3.82      6.9±7.92ms        ? ?/sec    1.00  1817.9±37.42µs        ? ?/sec
get path Vec2(356.0, 166.0)    1.00  1736.4±42.28µs        ? ?/sec    1.12  1937.9±32.68µs        ? ?/sec
get path Vec2(468.0, 584.0)    9.71     13.8±2.02ms        ? ?/sec    1.00  1417.5±44.26µs        ? ?/sec
get path Vec2(512.0, 170.0)    1.44      2.5±1.00ms        ? ?/sec    1.00  1765.4±49.34µs        ? ?/sec
get path Vec2(611.0, 658.0)    1.27      2.5±0.70ms        ? ?/sec    1.00  1979.8±15.14µs        ? ?/sec
get path Vec2(827.0, 678.0)    1.00  1909.9±35.95µs        ? ?/sec    1.07      2.1±0.02ms        ? ?/sec
get path Vec2(993.0, 290.0)    1.00      3.3±0.02ms        ? ?/sec    1.15      3.8±0.07ms        ? ?/sec

```


Also i don't know why but in execution times for first path there are large differences