@PTFOPlayer 2025-01-21
I did some more benchmarking, it seems like on "get path" benches there is some regression. For now i'm closing this PR and i'll reopen it when it is fixed